server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx>=0.28.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9
google-genai==1.14.0
pydub==0.25.1
